+++ http/src/http.mli
@@ -202,8 +202,8 @@ module Header : sig
       name [k] and value [v]. *)
 
   val add : t -> string -> string -> t
-  (** [add h k v] adds the header name [k] and it associated value [v] at the
-      end of header list [h]. *)
+  (** [add h k v] adds the header name [k] and its associated value [v] at the
+      front of header list [h]. *)
 
   val add_list : t -> (string * string) list -> t
   (** [add_list h l] adds in order all header pairs contained in [l] to the
@@ -366,6 +366,15 @@ module Header : sig
     val caseless_equal : string -> string -> bool
     (** [caseless_equal a b] must be equivalent to
         [String.equal (String.lowercase_ascii a) (String.lowercase_ascii b)]. *)
+
+    val first : t -> (string * string) option
+    (** [first t] is [Some (hdr_name, hdr_value)], which represents the first
+        header in headers list [t]. It is [None] if [t] is empty. *)
+
+    val move_to_front : t -> string -> t
+    (** [move_to_front t hdr_name] is [t] with header name [hdr_name] moved to
+        the front of the headers list [t]. If the header doesn't exist in [t] or
+        the header is already at the front, then [t] is unchanged. *)
   end
 end