+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- cohttp-eio: ensure "Host" header is the first header in http client requests (bikallem #939)
 - cohttp-eio: add TE header in client. Check TE header is server (bikallem #941)
 - cohttp-eio: add User-Agent header to request from Client (bikallem #940)
 - cohttp-eio: add Content-Length header to request/response (bikallem #929)